requirements.txt
streamlit==1.28.0
numpy==1.24.3
librosa==0.10.1
matplotlib==3.7.2
plotly==5.18.0
pandas==2.0.3
pillow==10.0.1
